pyproject.toml
[project]
name = "sharks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",
	"httpx>=0.27.0",
	"python-dotenv>=1.0.1",
	"pydantic>=2.8.0",
]
